Improvements and bug fixing.
The problem with the code was that it was using std::pow and std::fmod which caused it not to work. This can be fixed by making our own function for it. I've used ChatGPT to generate said function out of laziness and it might have some room for improvement.